TW-249: Find deadcode. -- References 'temple/assets' -> 'temple/front'
