+++ src/app/pages/ManageAssets.tsx
@@ -12,8 +12,8 @@ import { ManageAssetsSelectors } from 'app/pages/ManageAssets.selectors';
 import { AssetIcon } from 'app/templates/AssetIcon';
 import SearchAssetField from 'app/templates/SearchAssetField';
 import { T, t } from 'lib/i18n/react';
-import { AssetTypesEnum, setTokenStatus, useFilteredAssets } from 'lib/temple/assets';
-import { useAccount, useChainId, useAssetMetadata, useAvailableAssets } from 'lib/temple/front';
+import { AssetTypesEnum, setTokenStatus } from 'lib/temple/assets';
+import { useAccount, useChainId, useAssetMetadata, useAvailableAssets, useFilteredAssets } from 'lib/temple/front';
 import { getAssetName, getAssetSymbol } from 'lib/temple/metadata';
 import { ITokenStatus, ITokenType } from 'lib/temple/repo';
 import { useConfirm } from 'lib/ui/dialog';
