+++ src/app/templates/SwapForm/SwapFormInput/SwapFormInput.tsx
@@ -5,14 +5,15 @@ import classNames from 'clsx';
 
 import { useFormAnalytics } from 'lib/analytics';
 import { t } from 'lib/i18n/react';
-import { AssetTypesEnum, toTokenSlug, useFilteredAssets } from 'lib/temple/assets';
+import { AssetTypesEnum, toTokenSlug } from 'lib/temple/assets';
 import {
   useAccount,
   useBalance,
   useAssetMetadata,
   useAvailableAssets,
   useGetTokenMetadata,
-  useOnBlock
+  useOnBlock,
+  useFilteredAssets
 } from 'lib/temple/front';
 import { EMPTY_ASSET_METADATA } from 'lib/temple/metadata';
 import Popper from 'lib/ui/Popper';
