+++ src/lib/temple/front/assets.ts
@@ -1,8 +1,9 @@
-import { useCallback, useEffect, useMemo, useRef } from 'react';
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
 import constate from 'constate';
 import deepEqual from 'fast-deep-equal';
 import Fuse from 'fuse.js';
+import { useDebounce } from 'use-debounce';
 import useForceUpdate from 'use-force-update';
 import { browser } from 'webextension-polyfill-ts';
 
@@ -16,7 +17,8 @@ import {
   fetchAllKnownFungibleTokenSlugs,
   fetchCollectibleTokens,
   fetchAllKnownCollectibleTokenSlugs,
-  isTokenDisplayed
+  isTokenDisplayed,
+  toTokenSlug
 } from 'lib/temple/assets';
 import { useNetwork } from 'lib/temple/front';
 import {
@@ -290,6 +292,27 @@ export const useAvailableAssets = (assetType: AssetTypesEnum) => {
 
   return { availableAssets, assetsStatuses, isLoading, mutate };
 };
+
+export function useFilteredAssets(assetSlugs: string[]) {
+  const allTokensBaseMetadata = useAllTokensBaseMetadata();
+
+  const [searchValue, setSearchValue] = useState('');
+  const [tokenId, setTokenId] = useState<number>();
+  const [searchValueDebounced] = useDebounce(tokenId ? toTokenSlug(searchValue, tokenId) : searchValue, 300);
+
+  const filteredAssets = useMemo(
+    () => searchAssets(searchValueDebounced, assetSlugs, allTokensBaseMetadata),
+    [searchValueDebounced, assetSlugs, allTokensBaseMetadata]
+  );
+
+  return {
+    filteredAssets,
+    searchValue,
+    setSearchValue,
+    tokenId,
+    setTokenId
+  };
+}
 
 export function searchAssets(
   searchValue: string,
